refactored bonus-product-utils.js
